+++ lib/common/utils/device_utils.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:flutter/foundation.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:android_id/android_id.dart';
 import 'package:device_info_plus/device_info_plus.dart';
@@ -23,5 +24,18 @@ class DeviceUtils {
   static Future<String> getAppVersion() async {
     var info = await PackageInfo.fromPlatform();
     return "${info.version} (${info.buildNumber})";
+  }
+
+  static Future<String> getAppSemanticVersion() async {
+    var info = await PackageInfo.fromPlatform();
+    return info.version;
+  }
+
+  /// Checks if the current platform is mobile and returns true if it is.
+  static bool isMobilePlatform() {
+    if (kIsWeb) {
+      return false;
+    }
+    return Platform.isAndroid || Platform.isIOS;
   }
 }