+++ lib/features/appConfiguration/domain/usecase/get_app_config_usecase.dart
@@ -1,0 +1,14 @@
+import 'package:zachranobed/features/appConfiguration/domain/repository/app_configuration_repository.dart';
+import 'package:zachranobed/features/forceupdate/domain/model/app_config.dart';
+
+/// Use case to get application configuration.
+class GetAppConfigUseCase {
+  final AppConfigurationRepository _repository;
+
+  GetAppConfigUseCase(this._repository);
+
+  /// Fetches available application configuration.
+  Future<AppConfig> invoke() async {
+    return await _repository.getAppConfig();
+  }
+}