+++ lib/features/appTerms/domain/set_newest_accepted_app_terms_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:zachranobed/common/logger/zo_logger.dart';
 import 'package:zachranobed/features/appTerms/domain/update_current_user_data_use_case.dart';
-import 'package:zachranobed/features/appConfiguration/data/get_last_app_terms_version_use_case.dart';
+import 'package:zachranobed/features/appConfiguration/domain/usecase/get_last_app_terms_version_use_case.dart';
 
 class SetNewestAcceptedAppTermsUseCase {
   final UpdateCurrentUserDataUseCase _updateCurrentUserDataUseCase;
@@ -14,7 +14,7 @@ class SetNewestAcceptedAppTermsUseCase {
     if (lastAppTermsVersion != null) {
       _updateCurrentUserDataUseCase.updateLastAcceptedAppTermsVersion(lastAppTermsVersion);
     } else {
-      ZOLogger.logMessage("Unable to set the last accepted app terms because the remote value is `${lastAppTermsVersion}`", isError: true);
+      ZOLogger.logMessage("Unable to set the last accepted app terms because the remote value is `$lastAppTermsVersion`", isError: true);
     }
   }
 }