+++ lib/routes/app_router.gr.dart
@@ -8,53 +8,55 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i17;
-import 'package:flutter/material.dart' as _i18;
+import 'package:auto_route/auto_route.dart' as _i18;
+import 'package:flutter/material.dart' as _i19;
 import 'package:zachranobed/features/activepair/presentation/change_active_pair_screen.dart'
     as _i3;
 import 'package:zachranobed/features/appTerms/presentation/app_terms_screen.dart'
     as _i1;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i6;
 import 'package:zachranobed/features/foodboxes/domain/model/box_movement.dart'
-    as _i19;
+    as _i20;
 import 'package:zachranobed/features/foodboxes/presentation/screen/box_movement_detail_screen.dart'
     as _i2;
+import 'package:zachranobed/features/forceupdate/presentation/force_update_screen.dart'
+    as _i8;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
-    as _i10;
+    as _i11;
 import 'package:zachranobed/features/menu/presentation/contacts_screen.dart'
     as _i5;
 import 'package:zachranobed/features/menu/presentation/menu_screen.dart'
-    as _i11;
+    as _i12;
 import 'package:zachranobed/features/offeredfood/domain/model/food_info.dart'
-    as _i21;
+    as _i22;
 import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart'
-    as _i20;
+    as _i21;
 import 'package:zachranobed/features/offeredfood/presentation/screens/donated_food_detail_screen.dart'
     as _i7;
 import 'package:zachranobed/ui/screens/change_password_screen.dart' as _i4;
-import 'package:zachranobed/ui/screens/forgot_password_screen.dart' as _i8;
-import 'package:zachranobed/ui/screens/home_screen.dart' as _i9;
-import 'package:zachranobed/ui/screens/offer_food_detail_screen.dart' as _i12;
-import 'package:zachranobed/ui/screens/offer_food_overview_screen.dart' as _i13;
-import 'package:zachranobed/ui/screens/offer_food_screen.dart' as _i14;
+import 'package:zachranobed/ui/screens/forgot_password_screen.dart' as _i9;
+import 'package:zachranobed/ui/screens/home_screen.dart' as _i10;
+import 'package:zachranobed/ui/screens/offer_food_detail_screen.dart' as _i13;
+import 'package:zachranobed/ui/screens/offer_food_overview_screen.dart' as _i14;
+import 'package:zachranobed/ui/screens/offer_food_screen.dart' as _i15;
 import 'package:zachranobed/ui/screens/order_shipping_of_boxes_screen.dart'
-    as _i15;
-import 'package:zachranobed/ui/screens/thank_you_screen.dart' as _i16;
+    as _i16;
+import 'package:zachranobed/ui/screens/thank_you_screen.dart' as _i17;
 
-abstract class $AppRouter extends _i17.RootStackRouter {
+abstract class $AppRouter extends _i18.RootStackRouter {
   $AppRouter({super.navigatorKey});
 
   @override
-  final Map<String, _i17.PageFactory> pagesMap = {
+  final Map<String, _i18.PageFactory> pagesMap = {
     AppTermsRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i1.AppTermsScreen(),
       );
     },
     BoxMovementDetailRoute.name: (routeData) {
       final args = routeData.argsAs<BoxMovementDetailRouteArgs>();
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: _i2.BoxMovementDetailScreen(
           key: args.key,
@@ -63,32 +65,32 @@ abstract class $AppRouter extends _i17.RootStackRouter {
       );
     },
     ChangeActivePairRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i3.ChangeActivePairScreen(),
       );
     },
     ChangePasswordRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i4.ChangePasswordScreen(),
       );
     },
     ContactsRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i5.ContactsScreen(),
       );
     },
     DebugRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: const _i6.DebugScreen(),
       );
     },
     DonatedFoodDetailRoute.name: (routeData) {
       final args = routeData.argsAs<DonatedFoodDetailRouteArgs>();
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
         child: _i7.DonatedFoodDetailScreen(
           key: args.key,
@@ -96,35 +98,41 @@ abstract class $AppRouter extends _i17.RootStackRouter {
         ),
       );
     },
-    ForgotPasswordRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+    ForceUpdateRoute.name: (routeData) {
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i8.ForgotPasswordScreen(),
+        child: const _i8.ForceUpdateScreen(),
+      );
+    },
+    ForgotPasswordRoute.name: (routeData) {
+      return _i18.AutoRoutePage<dynamic>(
+        routeData: routeData,
+        child: const _i9.ForgotPasswordScreen(),
       );
     },
     HomeRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i9.HomeScreen(),
+        child: const _i10.HomeScreen(),
       );
     },
     LoginRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i10.LoginScreen(),
+        child: const _i11.LoginScreen(),
       );
     },
     MenuRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i11.MenuScreen(),
+        child: const _i12.MenuScreen(),
       );
     },
     OfferFoodDetailRoute.name: (routeData) {
       final args = routeData.argsAs<OfferFoodDetailRouteArgs>();
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: _i12.OfferFoodDetailScreen(
+        child: _i13.OfferFoodDetailScreen(
           key: args.key,
           editedFoodInfo: args.editedFoodInfo,
           allFoodInfos: args.allFoodInfos,
@@ -133,31 +141,31 @@ abstract class $AppRouter extends _i17.RootStackRouter {
     },
     OfferFoodOverviewRoute.name: (routeData) {
       final args = routeData.argsAs<OfferFoodOverviewRouteArgs>();
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: _i13.OfferFoodOverviewScreen(
+        child: _i14.OfferFoodOverviewScreen(
           key: args.key,
           foodInfos: args.foodInfos,
         ),
       );
     },
     OfferFoodRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i14.OfferFoodScreen(),
+        child: const _i15.OfferFoodScreen(),
       );
     },
     OrderShippingOfBoxesRoute.name: (routeData) {
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: const _i15.OrderShippingOfBoxesScreen(),
+        child: const _i16.OrderShippingOfBoxesScreen(),
       );
     },
     ThankYouRoute.name: (routeData) {
       final args = routeData.argsAs<ThankYouRouteArgs>();
-      return _i17.AutoRoutePage<dynamic>(
+      return _i18.AutoRoutePage<dynamic>(
         routeData: routeData,
-        child: _i16.ThankYouScreen(
+        child: _i17.ThankYouScreen(
           key: args.key,
           isSuccess: args.isSuccess,
           message: args.message,
@@ -169,8 +177,8 @@ abstract class $AppRouter extends _i17.RootStackRouter {
 
 /// generated route for
 /// [_i1.AppTermsScreen]
-class AppTermsRoute extends _i17.PageRouteInfo<void> {
-  const AppTermsRoute({List<_i17.PageRouteInfo>? children})
+class AppTermsRoute extends _i18.PageRouteInfo<void> {
+  const AppTermsRoute({List<_i18.PageRouteInfo>? children})
       : super(
           AppTermsRoute.name,
           initialChildren: children,
@@ -178,17 +186,17 @@ class AppTermsRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'AppTermsRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i2.BoxMovementDetailScreen]
 class BoxMovementDetailRoute
-    extends _i17.PageRouteInfo<BoxMovementDetailRouteArgs> {
+    extends _i18.PageRouteInfo<BoxMovementDetailRouteArgs> {
   BoxMovementDetailRoute({
-    _i18.Key? key,
-    required _i19.BoxMovement boxMovement,
-    List<_i17.PageRouteInfo>? children,
+    _i19.Key? key,
+    required _i20.BoxMovement boxMovement,
+    List<_i18.PageRouteInfo>? children,
   }) : super(
           BoxMovementDetailRoute.name,
           args: BoxMovementDetailRouteArgs(
@@ -200,8 +208,8 @@ class BoxMovementDetailRoute
 
   static const String name = 'BoxMovementDetailRoute';
 
-  static const _i17.PageInfo<BoxMovementDetailRouteArgs> page =
-      _i17.PageInfo<BoxMovementDetailRouteArgs>(name);
+  static const _i18.PageInfo<BoxMovementDetailRouteArgs> page =
+      _i18.PageInfo<BoxMovementDetailRouteArgs>(name);
 }
 
 class BoxMovementDetailRouteArgs {
@@ -210,9 +218,9 @@ class BoxMovementDetailRouteArgs {
     required this.boxMovement,
   });
 
-  final _i18.Key? key;
+  final _i19.Key? key;
 
-  final _i19.BoxMovement boxMovement;
+  final _i20.BoxMovement boxMovement;
 
   @override
   String toString() {
@@ -222,8 +230,8 @@ class BoxMovementDetailRouteArgs {
 
 /// generated route for
 /// [_i3.ChangeActivePairScreen]
-class ChangeActivePairRoute extends _i17.PageRouteInfo<void> {
-  const ChangeActivePairRoute({List<_i17.PageRouteInfo>? children})
+class ChangeActivePairRoute extends _i18.PageRouteInfo<void> {
+  const ChangeActivePairRoute({List<_i18.PageRouteInfo>? children})
       : super(
           ChangeActivePairRoute.name,
           initialChildren: children,
@@ -231,13 +239,13 @@ class ChangeActivePairRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'ChangeActivePairRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i4.ChangePasswordScreen]
-class ChangePasswordRoute extends _i17.PageRouteInfo<void> {
-  const ChangePasswordRoute({List<_i17.PageRouteInfo>? children})
+class ChangePasswordRoute extends _i18.PageRouteInfo<void> {
+  const ChangePasswordRoute({List<_i18.PageRouteInfo>? children})
       : super(
           ChangePasswordRoute.name,
           initialChildren: children,
@@ -245,13 +253,13 @@ class ChangePasswordRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'ChangePasswordRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i5.ContactsScreen]
-class ContactsRoute extends _i17.PageRouteInfo<void> {
-  const ContactsRoute({List<_i17.PageRouteInfo>? children})
+class ContactsRoute extends _i18.PageRouteInfo<void> {
+  const ContactsRoute({List<_i18.PageRouteInfo>? children})
       : super(
           ContactsRoute.name,
           initialChildren: children,
@@ -259,13 +267,13 @@ class ContactsRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'ContactsRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i6.DebugScreen]
-class DebugRoute extends _i17.PageRouteInfo<void> {
-  const DebugRoute({List<_i17.PageRouteInfo>? children})
+class DebugRoute extends _i18.PageRouteInfo<void> {
+  const DebugRoute({List<_i18.PageRouteInfo>? children})
       : super(
           DebugRoute.name,
           initialChildren: children,
@@ -273,17 +281,17 @@ class DebugRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'DebugRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
 /// [_i7.DonatedFoodDetailScreen]
 class DonatedFoodDetailRoute
-    extends _i17.PageRouteInfo<DonatedFoodDetailRouteArgs> {
+    extends _i18.PageRouteInfo<DonatedFoodDetailRouteArgs> {
   DonatedFoodDetailRoute({
-    _i18.Key? key,
-    required _i20.OfferedFood offeredFood,
-    List<_i17.PageRouteInfo>? children,
+    _i19.Key? key,
+    required _i21.OfferedFood offeredFood,
+    List<_i18.PageRouteInfo>? children,
   }) : super(
           DonatedFoodDetailRoute.name,
           args: DonatedFoodDetailRouteArgs(
@@ -295,8 +303,8 @@ class DonatedFoodDetailRoute
 
   static const String name = 'DonatedFoodDetailRoute';
 
-  static const _i17.PageInfo<DonatedFoodDetailRouteArgs> page =
-      _i17.PageInfo<DonatedFoodDetailRouteArgs>(name);
+  static const _i18.PageInfo<DonatedFoodDetailRouteArgs> page =
+      _i18.PageInfo<DonatedFoodDetailRouteArgs>(name);
 }
 
 class DonatedFoodDetailRouteArgs {
@@ -305,9 +313,9 @@ class DonatedFoodDetailRouteArgs {
     required this.offeredFood,
   });
 
-  final _i18.Key? key;
+  final _i19.Key? key;
 
-  final _i20.OfferedFood offeredFood;
+  final _i21.OfferedFood offeredFood;
 
   @override
   String toString() {
@@ -316,9 +324,23 @@ class DonatedFoodDetailRouteArgs {
 }
 
 /// generated route for
-/// [_i8.ForgotPasswordScreen]
-class ForgotPasswordRoute extends _i17.PageRouteInfo<void> {
-  const ForgotPasswordRoute({List<_i17.PageRouteInfo>? children})
+/// [_i8.ForceUpdateScreen]
+class ForceUpdateRoute extends _i18.PageRouteInfo<void> {
+  const ForceUpdateRoute({List<_i18.PageRouteInfo>? children})
+      : super(
+          ForceUpdateRoute.name,
+          initialChildren: children,
+        );
+
+  static const String name = 'ForceUpdateRoute';
+
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
+}
+
+/// generated route for
+/// [_i9.ForgotPasswordScreen]
+class ForgotPasswordRoute extends _i18.PageRouteInfo<void> {
+  const ForgotPasswordRoute({List<_i18.PageRouteInfo>? children})
       : super(
           ForgotPasswordRoute.name,
           initialChildren: children,
@@ -326,13 +348,13 @@ class ForgotPasswordRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'ForgotPasswordRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i9.HomeScreen]
-class HomeRoute extends _i17.PageRouteInfo<void> {
-  const HomeRoute({List<_i17.PageRouteInfo>? children})
+/// [_i10.HomeScreen]
+class HomeRoute extends _i18.PageRouteInfo<void> {
+  const HomeRoute({List<_i18.PageRouteInfo>? children})
       : super(
           HomeRoute.name,
           initialChildren: children,
@@ -340,13 +362,13 @@ class HomeRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'HomeRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i10.LoginScreen]
-class LoginRoute extends _i17.PageRouteInfo<void> {
-  const LoginRoute({List<_i17.PageRouteInfo>? children})
+/// [_i11.LoginScreen]
+class LoginRoute extends _i18.PageRouteInfo<void> {
+  const LoginRoute({List<_i18.PageRouteInfo>? children})
       : super(
           LoginRoute.name,
           initialChildren: children,
@@ -354,13 +376,13 @@ class LoginRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'LoginRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i11.MenuScreen]
-class MenuRoute extends _i17.PageRouteInfo<void> {
-  const MenuRoute({List<_i17.PageRouteInfo>? children})
+/// [_i12.MenuScreen]
+class MenuRoute extends _i18.PageRouteInfo<void> {
+  const MenuRoute({List<_i18.PageRouteInfo>? children})
       : super(
           MenuRoute.name,
           initialChildren: children,
@@ -368,18 +390,18 @@ class MenuRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'MenuRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i12.OfferFoodDetailScreen]
+/// [_i13.OfferFoodDetailScreen]
 class OfferFoodDetailRoute
-    extends _i17.PageRouteInfo<OfferFoodDetailRouteArgs> {
+    extends _i18.PageRouteInfo<OfferFoodDetailRouteArgs> {
   OfferFoodDetailRoute({
-    _i18.Key? key,
-    required _i21.FoodInfo editedFoodInfo,
-    required List<_i21.FoodInfo> allFoodInfos,
-    List<_i17.PageRouteInfo>? children,
+    _i19.Key? key,
+    required _i22.FoodInfo editedFoodInfo,
+    required List<_i22.FoodInfo> allFoodInfos,
+    List<_i18.PageRouteInfo>? children,
   }) : super(
           OfferFoodDetailRoute.name,
           args: OfferFoodDetailRouteArgs(
@@ -392,8 +414,8 @@ class OfferFoodDetailRoute
 
   static const String name = 'OfferFoodDetailRoute';
 
-  static const _i17.PageInfo<OfferFoodDetailRouteArgs> page =
-      _i17.PageInfo<OfferFoodDetailRouteArgs>(name);
+  static const _i18.PageInfo<OfferFoodDetailRouteArgs> page =
+      _i18.PageInfo<OfferFoodDetailRouteArgs>(name);
 }
 
 class OfferFoodDetailRouteArgs {
@@ -403,11 +425,11 @@ class OfferFoodDetailRouteArgs {
     required this.allFoodInfos,
   });
 
-  final _i18.Key? key;
+  final _i19.Key? key;
 
-  final _i21.FoodInfo editedFoodInfo;
+  final _i22.FoodInfo editedFoodInfo;
 
-  final List<_i21.FoodInfo> allFoodInfos;
+  final List<_i22.FoodInfo> allFoodInfos;
 
   @override
   String toString() {
@@ -416,13 +438,13 @@ class OfferFoodDetailRouteArgs {
 }
 
 /// generated route for
-/// [_i13.OfferFoodOverviewScreen]
+/// [_i14.OfferFoodOverviewScreen]
 class OfferFoodOverviewRoute
-    extends _i17.PageRouteInfo<OfferFoodOverviewRouteArgs> {
+    extends _i18.PageRouteInfo<OfferFoodOverviewRouteArgs> {
   OfferFoodOverviewRoute({
-    _i18.Key? key,
-    required List<_i21.FoodInfo> foodInfos,
-    List<_i17.PageRouteInfo>? children,
+    _i19.Key? key,
+    required List<_i22.FoodInfo> foodInfos,
+    List<_i18.PageRouteInfo>? children,
   }) : super(
           OfferFoodOverviewRoute.name,
           args: OfferFoodOverviewRouteArgs(
@@ -434,8 +456,8 @@ class OfferFoodOverviewRoute
 
   static const String name = 'OfferFoodOverviewRoute';
 
-  static const _i17.PageInfo<OfferFoodOverviewRouteArgs> page =
-      _i17.PageInfo<OfferFoodOverviewRouteArgs>(name);
+  static const _i18.PageInfo<OfferFoodOverviewRouteArgs> page =
+      _i18.PageInfo<OfferFoodOverviewRouteArgs>(name);
 }
 
 class OfferFoodOverviewRouteArgs {
@@ -444,9 +466,9 @@ class OfferFoodOverviewRouteArgs {
     required this.foodInfos,
   });
 
-  final _i18.Key? key;
+  final _i19.Key? key;
 
-  final List<_i21.FoodInfo> foodInfos;
+  final List<_i22.FoodInfo> foodInfos;
 
   @override
   String toString() {
@@ -455,9 +477,9 @@ class OfferFoodOverviewRouteArgs {
 }
 
 /// generated route for
-/// [_i14.OfferFoodScreen]
-class OfferFoodRoute extends _i17.PageRouteInfo<void> {
-  const OfferFoodRoute({List<_i17.PageRouteInfo>? children})
+/// [_i15.OfferFoodScreen]
+class OfferFoodRoute extends _i18.PageRouteInfo<void> {
+  const OfferFoodRoute({List<_i18.PageRouteInfo>? children})
       : super(
           OfferFoodRoute.name,
           initialChildren: children,
@@ -465,13 +487,13 @@ class OfferFoodRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'OfferFoodRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i15.OrderShippingOfBoxesScreen]
-class OrderShippingOfBoxesRoute extends _i17.PageRouteInfo<void> {
-  const OrderShippingOfBoxesRoute({List<_i17.PageRouteInfo>? children})
+/// [_i16.OrderShippingOfBoxesScreen]
+class OrderShippingOfBoxesRoute extends _i18.PageRouteInfo<void> {
+  const OrderShippingOfBoxesRoute({List<_i18.PageRouteInfo>? children})
       : super(
           OrderShippingOfBoxesRoute.name,
           initialChildren: children,
@@ -479,17 +501,17 @@ class OrderShippingOfBoxesRoute extends _i17.PageRouteInfo<void> {
 
   static const String name = 'OrderShippingOfBoxesRoute';
 
-  static const _i17.PageInfo<void> page = _i17.PageInfo<void>(name);
+  static const _i18.PageInfo<void> page = _i18.PageInfo<void>(name);
 }
 
 /// generated route for
-/// [_i16.ThankYouScreen]
-class ThankYouRoute extends _i17.PageRouteInfo<ThankYouRouteArgs> {
+/// [_i17.ThankYouScreen]
+class ThankYouRoute extends _i18.PageRouteInfo<ThankYouRouteArgs> {
   ThankYouRoute({
-    _i18.Key? key,
+    _i19.Key? key,
     required bool isSuccess,
     required String message,
-    List<_i17.PageRouteInfo>? children,
+    List<_i18.PageRouteInfo>? children,
   }) : super(
           ThankYouRoute.name,
           args: ThankYouRouteArgs(
@@ -502,8 +524,8 @@ class ThankYouRoute extends _i17.PageRouteInfo<ThankYouRouteArgs> {
 
   static const String name = 'ThankYouRoute';
 
-  static const _i17.PageInfo<ThankYouRouteArgs> page =
-      _i17.PageInfo<ThankYouRouteArgs>(name);
+  static const _i18.PageInfo<ThankYouRouteArgs> page =
+      _i18.PageInfo<ThankYouRouteArgs>(name);
 }
 
 class ThankYouRouteArgs {
@@ -513,7 +535,7 @@ class ThankYouRouteArgs {
     required this.message,
   });
 
-  final _i18.Key? key;
+  final _i19.Key? key;
 
   final bool isSuccess;