+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is a changelog for **ZachranObed** application.
 ### Fixed
 
 ### Added
+- **ZOB-318** Add ForceUpdate screen functionality.
 
 ### Fixed